Use the right naming for toml-rb.
To get this code working with the latest version of toml-rb, we have to
make some minor changes. First of all, "require 'toml'" will pull in
another TOML library with a different API. I also had to make use of
TomlRB.dump() instead of TOML.dump() to get this code to work.